+++ src/content/dashboards/Reports/CalendarTodayReservation.tsx
@@ -1,16 +1,12 @@
-import { useState, useRef, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import FullCalendar, { EventDropArg } from '@fullcalendar/react';
+import FullCalendar from '@fullcalendar/react';
 import resourceTimelinePlugin from '@fullcalendar/resource-timeline';
-import interactionPlugin, {
-  EventResizeDoneArg
-} from '@fullcalendar/interaction';
+import interactionPlugin from '@fullcalendar/interaction';
 import jaLocale from '@fullcalendar/core/locales/ja';
 import { Box, Card, styled } from '@mui/material';
 import { setHours, setMinutes, setSeconds } from 'date-fns';
-import { useOrderCalendarState } from 'src/contexts/OrderCalendarContext';
-import { useOrderCalendar } from 'src/hooks/useOrderCalendar';
-import { useOrder } from 'src/hooks/useOrder';
+import { useCalendarTodayReservationState } from './store';
 
 const FullCalendarWrapper = styled(Box)(
   ({ theme }) => `
@@ -23,41 +19,16 @@ const FullCalendarWrapper = styled(Box)(
 
 const CalendarTodayReservation = () => {
   const { t }: { t: any } = useTranslation();
-  const calendarRef = useRef<FullCalendar | null>(null);
-  const [date, setDate] = useState<Date>(new Date());
-  const { getOrderCalendarTimeline, putOrderCalendar } = useOrderCalendar();
-  const { events, resources } = useOrderCalendarState();
-  const { putOrderTime } = useOrder();
+  const {
+    calendarRef,
+    date,
+    events,
+    resources,
 
-  const handleEventResize = ({ event }: EventResizeDoneArg): void => {
-    try {
-      putOrderTime(parseInt(event.id), {
-        dateOfVisit: event.start,
-        dateOfExit: event.end
-      });
-      putOrderCalendar(event.id, {
-        start: event.start,
-        end: event.end
-      });
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const handleEventDrop = ({ event }: EventDropArg): void => {
-    try {
-      putOrderTime(parseInt(event.id), {
-        dateOfVisit: event.start,
-        dateOfExit: event.end
-      });
-      putOrderCalendar(event.id, {
-        start: event.start,
-        end: event.end
-      });
-    } catch (err) {
-      console.error(err);
-    }
-  };
+    getOrderCalendarTimeline,
+    handleEventResize,
+    handleEventDrop
+  } = useCalendarTodayReservationState();
 
   useEffect(() => {
     const startDate: Date = setSeconds(setMinutes(setHours(date, 0), 0), 0);
