+++ src/content/dashboards/Reports/TodayReservationList.tsx
@@ -19,6 +19,7 @@ import Scrollbar from 'src/components/Scrollbar';
 import { useOrder } from 'src/hooks/useOrder';
 import { useOrderState } from 'src/contexts/OrderContext';
 import UpdateOrderForm from './UpdateOrderForm';
+import { useTodayReservationListState } from './store';
 
 const LabelComplete = styled(Box)(
   ({ theme }) => `
@@ -76,22 +77,20 @@ const getOrderStatusLabel = (status: Status) => {
 
 const TodayReservationList = () => {
   const { t }: { t: any } = useTranslation();
-  const { orders } = useOrderState();
-  const { getTodayReservationOrders } = useOrder();
-  const [page, setPage] = useState(0);
-  const [limit, setLimit] = useState(10);
-  const [query, setQuery] = useState('');
+  const {
+    orders,
+    page,
+    limit,
+    query,
+    orderInfo,
+    open,
 
-  const [orderInfo, setOrderInfo] =
-    useState<{ orderId: number; name: string; phoneNumber: string }>();
-  const [open, setOpen] = useState(false);
-  const handleOpen = useCallback(() => setOpen(true), []);
-  const handleClose = useCallback(() => setOpen(false), []);
-
-  const handleQueryChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    event.persist();
-    setQuery(event.target.value);
-  };
+    getTodayReservationOrders,
+    handleOpen,
+    handleClose,
+    handleQueryChange,
+    setOrderInfo
+  } = useTodayReservationListState();
 
   useEffect(() => {
     getTodayReservationOrders(page, limit);
